refactor: Wasm ビルドから OS 依存コードを物理的に排除するため JSON ロード処理を分離
