+++ split-pass-api/internal/infra/fareio/json_test.go
@@ -1,4 +1,7 @@
+//go:build !js || !wasm
+
 package fareio_test
+
 
 import (
 	"os"
